User test

+++ src/main/java/com/example/kromannreumert/user/dto/UserResponseDTO.java
@@ -4,5 +4,5 @@ import com.example.kromannreumert.user.entity.Role;
 
 import java.util.Date;
 
-public record UserResponseDTO(int userId, String username, String name, String email, Date createdDate, Role role) {
+public record UserResponseDTO(long userId, String username, String name, String email, Date createdDate, Role role) {
 }
